made the CanDumpConverter a tool that is part of the lib
